Add jQuery circle progress plugin and demo page
The circle progress plugin, a dependent HTML demo page, and composer.json have been added. The jQuery plugin allows for the creation and control of circle progress on the page. The demo page is also set up to demonstrate the usage of the plugin. The composer.json file further outlines the dependencies needed for these files to work, such as jQuery and Bootstrap.